requirements.txt
markdown2
werkzeug
flask
python-mysql-connector